no arguments functionality
did this on in vscode so I had some trouble running ctrl d which would set the end of file when inputting from stdin which then created the reverse file. made the argument selection and rfile and wfile point to the correct locations with if and else statements. added more error handling